Added LOGIN MACRO and Update CW Speed/Status when using CWENABLE Macros.
This chnage allows a function key to be programmed to chnage operators. This is makes it easier to find for operators. The secnd is a bug where the CW status does not chnage until CW is sent. This makes it confusing after CWDISABLE is used and the CW speed is still set.