Commented out Git update messages so code will build
I was not certain which code was supposed to be there, but I updated the
LOGSUB1 file to comment out what I thought was wrong. The code in
Cryptic message needs to be reviewed to see what was desired by Howie. I
also fixed the project file options to remove d:\newsrc. Delphi allows
relative directories in the search options. This way no matter who
downloads the code (and to whatever directory), the code will compile
properly.