This fixes issue 88. In this fix, I added IC7300 as a valid radio type but also added the IC7100, IC7850 and IC7851. Additionally, I fixed some code that checked if the radio was Icom by testing an enumeration range of IC706..IC7800. The problem is that the Icom radios table grew above and below that. I updated this code to test IC78..IC9100. Hopefully, there will not be something above the 9100 for some time. The fundamental issue is that these tests check for the radio model instead of the RadioType of rtIcom. That would have handled it quite well.
With the above changes, I had to fix up some code in the CWBYCAT code to add the new types. I added all the radios that support CWByCat to a set defined in InitRadios in LogRadio.pas. If the radio does NOT support CWByCat, we do not do any of those commands. I toyed with the idea of not allow sCWByCAT to be set if the radio did not support it, but I did not want to change a config option under the user's feet so to speak.