Added contactdelete record
Added the contactdelete record. This record is sent when the user selected Alt-Y to delete last QSO. If the user then selects Alt-Y again to restore, the LogRecord is resent. This also changed the editQSO so if the user edits the QSO in the form, the delete and then re-add is sent. This addresses issue #165 .