Added code to send the Score broadcast every 250 seconds
To accurately send the Scores broadcast for WRTC, the procedure CalculateTotals has to be run. As that reads the log, I do not want to do it that often. This is because we need the QSO points by band.

This change added code to check a time and see if 250 seconds has elapsed. That is checked after each QSO so the station has to make a QSO to send it. I guess I should also do that on program shutdown too.