Common routine to generate My Exchange
I added a routine to create the MY Exchange for STX_STRING. This could also be used to generate the CABRILLO Sent exchange but I did not make that change yet. I have to create a testing framework for these so I can verify I hit all the Exchange cases but since it is ADIF export, that is OK to test more extenively in the field. I have a bit more to do but may call this ready to integrate into the repo. More testing to do though...