Lots of changes
Changed RSTReceived and RSTSent to a smallInt from a Word. Word is unsigned and FT8 and FT4 can have negative signal reports. This required changes in several places. In CAB export, it assumed if greater thaqn 99 the length is three without regard to sign. The EditQSO dialog required updating.

I also added the <app> field to the UDP packets. That closes #310

Another change is to update the mode to use the ExtendedMode string instead of just digital. This way FT8 and FT4 contacts show up properly.

CATEGORY-OPERATOR for Cabrillo is automatically generated from the list of operators in the log. This closes #357.