Put extendedModes in Qso Edit
Since we now have extended modes (CW, SSB, FT8, FT4, RTTY, etc), we show the extended modes in the mode selection.  If a user changed the mode of a contact, it updates the mode and extendedMode properly.

The idea is this is transparent tothe user. To minimize changes, the user using CW or SSB just sees those modes as normal. But FT8 and FT4 users, see the modes as they need.

I think I forgot int he prior committ, that I also fixed the Cabrillo format for Winter Field Day. That closes #370.