Found a bug in ADIF export for non extended
If the Extended mode was not set (not an imported contact), the MODe string emitted to ADIF was always CW. I added an eNoMode to the  ExtendedModes and cleared it on ClearContestExchange. This was a bug abnd testing #44 found it. This was not actually an issue any longer in the current code and this was a new bug introduced by by code--it did let yous set it to FM just fine though. So this closes #44.