More work for Split mode in WSJT-X
Split works on the K3 but not yet the Icom 7300. THis has to do with the way the program treats the Icom VFOs. It still holds on to the old way you had to deal with Icom (non-addressable VFOB). That works now so the code can be update for newer Icom radios.