WSJT-X fixes after Batavia  some others
This fixed the following issues I logged as working Batavia:

When new contact, the mults  in the log entry were not updates.

Added the QSO number to the WSJT-X contact (although this still has a minor issue).

Batavia prefix mult was not being added.

Cabrillo entry was not displaying properly for Digital contest. COnverted SNR to RST.

When we call a station, put the call int he call window so we can see where it is.

Also fixed an issue when you use enter a frequency into the call window to reset the QSO and Mults needed on the main log window. This closes #19 .

Also added more logger debugger displays.