When we call a station, display the proper grid and heading on the main log window
SInce we know the grid of the station we are calling, display the proper one. The program was figuring out the center of the grid when we put a call in the call window but that was a guess (center of call area I presume). Since WSJT-X knows the grid with certaintly, set it to that with DisplayBeamHeading...