Fixed the issue plus a few
The user option determines of we send highlight commands ot WSJT-X. Defaults to true.
Added email to WWDIGI contest.
Overloaded the TRACE and DEBUG method so you can use a Format like statement. This prevents doing needless string concat if debug is not set. i.e., logger.debug('The value of myString is %s',[myString]); When this is called, if debug is not set, the string is never formatted. This makes non-debug mode very fast.
Fixed the Yaesu FT891 split command of ST;
Changes to DUPE or MULT color scheme is sent ot WSJTX when changed (was only at start-up before).
In WSJT-X, removed some unused variables.