Fixed exporting the state
I fixed exporting the state bu adding more intelligence to the process. The code was looking at DomesticMults but that was wrong. I changed this to look at the contestsArray[contest].DF field to see if it was arrlsect and also to only emit the STATE ADIF field if we can really calculate a valid state.