FT4 and FT8 mode fixed and CAB in RTTY contest
This closes issue #457 and closes #466 . The fix to make the mode show up properly in the CAB and ADIF file was handled in PostUnit. The real issue here is that we we Digital for too much. I have a branch I have to finish testing that makes Digital and RTTY seperate Modes. So until that is all done, doing the fixup in Post makes the most sense.