Fix to Winter Field Day scoring and Cab generation
I use the power version to calculate the score so th euser onlyneeds to add in their bonus points. I also updated the QSOPoint calculation to include Phone and FM as modes that are one point. Previously it was just phone so FM ended up with 2 points. Field Day had the same bug.