Update LOGRADIO.PAS
The code to trace output was not called if the length was 3, 6 or 7. This may have excluded needed strings for debug.
Note we poll far to often. Consider reducing poll frequency.

Also, only tracing for Icom radios is in hex. Most other radios are in text format (some older radios may not be).