Added grid-locator to header
Added GRID-LOCATOR to the header as that is a valid field. This closes #599.

Also, cleared the Cabrillo Exchange fields before each contact is written. There was random data appearing in the CAB file.

Also, fixed an outstanding issue in the same place where the operators array sorted from the log is placed in the order of the number of contacts made. This seems more based on merit. The only place this really applies is for a multi-op and when a certificate is received. The calls are typically listed in the same order as the Operators field. Minor issue.