Added M class and MX and fixed a few related items.
Changed ADIF name to WFD.
Added M class.
Added support for receiving MX as a section for Mexico.
Also fixed #484.

I also found an error in TREE LooksLikeAGrid. Perhaps something was not committed. The line pAE := ActiveExchange was of the wrong type but it was never again referenced so I commented it out.

In fixing #484, I also reformatted the band/mode so the columns aligned better. It would be good to check ion a few other contests too.