Initial commit
Fixed #611 where the mode sent to UDP was SSB instead of LSB or USB.

Fixed latent defect where if DXLab's Commander was running, TR4W wqould receive an exception because of a conflict with the tcp port that Commander listens to. The port is the same as the one TR4W listens to for Commander emulation. I now handle the exception and log a message.