Add POTA
Misc fixes and add initial support for POTA. It is not a contest but it was a request from a local club member. I added support for park to park in multiple parks too. The exchange is K-1234/K-9999 and it creates two log entries. The ADIF is exported properly. Note the 2-fer park stuff is not quite done.
I removed a bunch of commeneted out code as it gets int he way when searching for other things.

I also changed the way the - key was used. Instead of an accelerator, I just search for it when someone presses it in the CallSign box.