Complete POTA
Fixed up the ADIF file for POTA.
Added a new Format in TF to handle 5 PChar parameters.
Allowed a blank exchange for POTA (default RST and no park)
Changed the directory name for the log for POTA to include the date as this is not a once a year contest
Updated the RES files again to remove the - as an accelerator (I'm not sure how that was overwritten when I refreshed my branch from MASTER).
Updated language constants to add a string for entering the POTA park upon new contest.