Completed code
This implements two new commands, RADIO ONE KEYER STOP BITS and RADIO TWO KEYER STOP BITS. This only needs to be set if there is an issue as it appears to work with 0 stop bits as opposed to the prior 2 that prevented the Yaesu serial ports from working. I also updated the HELP file descriptions. This closes #678