Fixed up UDP band issue
Fixed an issue where the band being sent in the UDP Contact record was shifted due to an array ordering issue.

Also fixed an issue where and if the band had a decimal (1.8, 3.5, 5.4), it was not localized as per the N1MM UDP documentation.

Also fixed a related issue where the DecimnalSeparator was set to . incorrectly when sending the whole log to UDP.

Added Try..Finally code around the kludge to reset the DecimalSeparator to ensure it is set back to the proper value for the locale.