Fix score posted to online scoreboard and UDP
The score sent to the online contest post was including the deleted contacts in the  total. This was incorrect. THe UDP sentfor score had the same issue. This was fixed as well. This closes #702.

Note I also moved some debug messages to the TRACE level to clean up the DEBUG output. This was mostly CWbyCAT related messages.