Add framework for more external loggers
The External Logger is now an array option with None, DXKeeper, ACLog and HRD.  I have not yet implemented the TCP send for ACLog and HRD but support is there to add it. Maybe someone will want to tackle that project if they use that log. UDP still works but this allows for programs that do not support direct UDP reception of QSO data (like DXKeeper). Note it would also be possible to send contacts logged into TR4W to N1MM. I'm not sure of the use case of that but it presents an intriguing option.