Fixed reset ports and other items in the  description
Made reset ports work for a networked K4. This is not quite done as the band is not getting set until you change bands I will work onthis in a future change so you do not even need to do the reset on a networked radio. This closes #737

Also updated the OpenURI function so this closes #731
.