Added code for 7760 and RegEx maintenance
Moved some code into includes to ensure TPerlRegEx is compiled with same version of code as pcre library. This should never change so it can just stay here.

Also put a newly compiled version of SCPFileUpload.dcu to account for using Indy 10. They had to be compiled with same version of Indy.