Added IC-7760
Added Icom IC-7760 and also fixed uSuperCheckPartialFileUpl;oad.dcu version again. This version was compiled with Indy 10. As long as another developer is using Indy 10 in their search path, it should match.

If not, email me as we have to find if there is an old version 9 of Indy hiding on the machine that has an error.