Add ID field support for contacts
This release generally adds an ID field to the contact record. To do this, it has to expand the log record to accommodate a 36 byte GUID. This will close #674. I also added code to add the SentExchange to the UDP Broadcasts--which will close #768.

Note this requires a new version of the TRSERVER as the log size changed. It also required bumping the log version number and allowing the log conversion routine to update old 1.6 logs to version 1.7.