Updates to ARRL-VHF, UDP Lookup
Fixed an issue where the exchange for ARRLVHF was incorrect in the ADIF message sent.

Added back the ZERO_NN in the ContestExchange record.

Fixed a typo in the XML UDP for IsClaimedQSO.

Added a LOokupInfoUDP with just the call as we do not yet have a CE record established.

Added a support function called IsValidMulticastIPAddress in preparation for multi-cast UDP for WSJT-X to get along with JTALERT and SpotCollector