Field Day and ARRL Section changes
FIxed issues with importing Field Day ADIF files.
Fixed an issue where if the shortcut section like WM was entered, we update that to WMA. Otherwise when outputting the section in the Cabrillo file, we would send WM.

Fixed issues importing DX stations to default to 1D and DX.

Added a helper function calls LooksLIkeAShortSection to see if that matches a section abbreviation.

Completly rewrote the routing that reads an ADIF file to handle multiple line records where a CRLF is in the middle of a field. This could happen on an address.