Add codes to allow VCL Forms
I found out the secret to adding standard Delphi VCL forms to the project. In my test, I can now bring up a form I created with the Delphi form editor. This allows me to create new forms for features as well as backfill existing forms like the BandMap for example or the Telnet window which have always been a problem when running as Windows forms.

It's a theory anyway...